backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
httpx==0.26.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
google-auth==2.34.0
requests==2.31.0
slowapi==0.1.9

